+++ app/src/main/java/ru/practicum/android/diploma/feature/filters/ui/IndustriesScreen.kt
@@ -1,15 +1,11 @@
 package ru.practicum.android.diploma.feature.filters.ui
 
-import android.media.Image
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.content.MediaType.Companion.Image
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.selection.selectable
@@ -24,12 +20,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kotlinx.coroutines.flow.MutableStateFlow
 import ru.practicum.android.diploma.R
@@ -48,6 +40,7 @@ fun IndustriesScreen(
     onBackClick: () -> Unit
 ) {
     val state by stateViewModel.collectAsStateWithLifecycle()
+    var text by remember { mutableStateOf("") }
 
     Scaffold(
         topBar = {
@@ -60,17 +53,21 @@ fun IndustriesScreen(
         Column(
             modifier = modifier.padding(paddingValues)
         ) {
+            SearchBar(
+                text = text,
+                onTextChange = { text = it }
+            )
             if (state.industries.isNotEmpty()) {
-                ShowIndustries(state.industries)
+                ShowContent(state.industries)
             } else {
-                ShowPlaceholder()
+                PlaceholderErrorListFetch()
             }
         }
     }
 }
 
 @Composable
-private fun ShowIndustries(industries: List<Industry>) {
+private fun ShowContent(industries: List<Industry>) {
     var selectedOption by remember { mutableStateOf<Industry?>(null) }
 
     LazyColumn(
@@ -82,7 +79,7 @@ private fun ShowIndustries(industries: List<Industry>) {
             Row(
                 modifier = Modifier
                     .fillMaxWidth()
-                    .height(60.dp)
+                    .height(AppDimensions.LabelActionListItem.itemHeight)
                     .selectable(
                         selected = (industry == selectedOption),
                         onClick = { selectedOption = industry }
@@ -107,32 +104,6 @@ private fun ShowIndustries(industries: List<Industry>) {
                 )
             }
         }
-    }
-}
-
-@Composable
-private fun ShowPlaceholder() {
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Image(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = AppDimensions.paddingMedium)
-                .padding(top = 122.dp),
-            painter = painterResource(R.drawable.error_list_fetch),
-            contentDescription = "error_list_fetch",
-            contentScale = ContentScale.Crop
-        )
-        Text(
-            modifier = Modifier
-                .width(268.dp)
-                .padding(top = AppDimensions.paddingMedium),
-            text = stringResource(R.string.error_list_fetch),
-            style = AppTypography.titleMedium,
-            textAlign = TextAlign.Center
-        )
     }
 }
