+++ app/src/main/java/ru/practicum/android/diploma/feature/filters/ui/SearchBar.kt
@@ -1,0 +1,83 @@
+package ru.practicum.android.diploma.feature.filters.ui
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import ru.practicum.android.diploma.R
+import ru.practicum.android.diploma.app.ui.theme.DiplomaTheme
+
+@Composable
+fun SearchBar(
+    text: String,
+    onTextChange: (String) -> Unit
+) {
+    OutlinedTextField(
+        value = text,
+        onValueChange = onTextChange,
+        placeholder = {
+            Text(
+                text = stringResource(R.string.hint_search_vacancy),
+                style = MaterialTheme.typography.bodyLarge,
+                color = MaterialTheme.colorScheme.outline
+            )
+        },
+        singleLine = true,
+        trailingIcon = {
+            if (text.isNotEmpty()) {
+                IconButton(onClick = { onTextChange("") }) {
+                    Icon(
+                        painterResource(R.drawable.ic_close_24),
+                        tint = MaterialTheme.colorScheme.onSecondary,
+                        contentDescription = null
+                    )
+                }
+            } else {
+                Icon(
+                    painterResource(R.drawable.ic_search_24),
+                    tint = MaterialTheme.colorScheme.onSecondary,
+                    contentDescription = null
+                )
+            }
+        },
+        shape = RoundedCornerShape(12.dp),
+        colors = TextFieldDefaults.colors(
+            focusedContainerColor = MaterialTheme.colorScheme.surfaceVariant,
+            unfocusedContainerColor = MaterialTheme.colorScheme.surfaceVariant,
+            cursorColor = MaterialTheme.colorScheme.primary,
+            focusedIndicatorColor = Color.Transparent,
+            unfocusedIndicatorColor = Color.Transparent,
+            focusedTextColor = MaterialTheme.colorScheme.onSecondary
+        ),
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp)
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun SearchBarPreview() {
+    val textState = remember { mutableStateOf("Введите запрос") }
+
+    DiplomaTheme {
+        SearchBar(
+            text = textState.value,
+            onTextChange = { textState.value = it },
+        )
+    }
+}
